Moved OrbitVersion into its own component
