+++ OrbitCore/CMakeLists.txt
@@ -23,7 +23,6 @@ target_sources(
          LinuxTracingBuffer.h
          OrbitModule.h
          OrbitProcess.h
-         OrbitVersion.h
          Params.h
          Path.h
          Pdb.h
@@ -95,11 +94,6 @@ target_compile_definitions(OrbitCore PUBLIC UNICODE)
 target_compile_definitions(OrbitCore PUBLIC _UNICODE)
 
 target_compile_features(OrbitCore PUBLIC cxx_std_11)
-
-include("${CMAKE_SOURCE_DIR}/cmake/version.cmake")
-GenerateVersionFile("${CMAKE_CURRENT_BINARY_DIR}/OrbitVersion.cpp"
-                    "${CMAKE_CURRENT_SOURCE_DIR}/OrbitVersion.cpp.in" OrbitCore)
-target_sources(OrbitCore PRIVATE "${CMAKE_CURRENT_BINARY_DIR}/OrbitVersion.cpp")
 
 add_executable(OrbitCoreTests)
